chore: components/host에 있던 컴포넌트를 components/specific으로 이동
