+++ frontend/src/components/specific/infoRow/InfoRow.styles.ts
