+++ frontend/src/components/specific/infoRow/InfoRow.tsx
