+++ frontend/src/components/specific/spaceCard/SpaceCard.styles.ts
