+++ frontend/src/components/specific/spaceCard/SpaceCard.tsx
