+++ frontend/src/pages/guest/workDetail/GuestWorkDetail.tsx
@@ -1,8 +1,8 @@
 import Button from '../../../components/@common/buttons/button/Button';
 import Footer from '../../../components/@common/footer/Footer';
 import { DividerLine } from '../../../styles/@common/DividerLine.styles';
-import * as C from '../../../styles/@common/WorkDetail.styles';
 import { mockWorkDetail } from '../../mockData';
+import * as C from '../../WorkDetail.common.styles';
 import * as S from './GuestWorkDetail.styles';
 
 const GuestWorkDetail = () => {
